cleanup Forwards.hpp
