+++ Indexing/ClauseVariantIndex.hpp
@@ -28,6 +28,7 @@ namespace Indexing {
 
 using namespace Lib;
 using namespace Kernel;
+class LiteralSubstitutionTree;
 
 class ClauseVariantIndex
 {
