+++ Inferences/TheoryInstAndSimp.hpp
@@ -24,6 +24,8 @@
 #include "Shell/Options.hpp"
 #include "SAT/Z3Interfacing.hpp"
 
+namespace Saturation { class Splitter; }
+
 namespace Inferences
 {
