+++ Kernel/Problem.hpp
@@ -80,15 +80,6 @@ public:
    */
   const TrivialPredicateMap& trivialPredicates() const { return _trivialPredicates; }
 
-  /**
-   * Always exactly one of the pair is non-zero, if the literal is specified,
-   * it must be ground.
-   */
-  typedef pair<Literal*,Clause*> BDDMeaningSpec;
-  typedef DHMap<unsigned, BDDMeaningSpec> BDDVarMeaningMap;
-  void addBDDVarMeaning(unsigned var, BDDMeaningSpec spec);
-  const BDDVarMeaningMap& getBDDVarMeanings() const { return _bddVarSpecs; }
-
   void addEliminatedFunction(unsigned func, Literal* definition);
   void addEliminatedPredicate(unsigned pred, Unit* definition);
   void addPartiallyEliminatedPredicate(unsigned pred, Unit* definition); 
@@ -207,7 +198,6 @@ private:
   bool _hadIncompleteTransformation;
 
   DHMap<unsigned,bool> _trivialPredicates;
-  BDDVarMeaningMap _bddVarSpecs;
 
   mutable bool _mayHaveEquality;
   mutable bool _mayHaveFormulas;
