+++ Lib/Stack.hpp
@@ -35,9 +35,6 @@ void swap(Lib::Stack<T>& s1, Lib::Stack<T>& s2);
 
 namespace Lib {
 
-template<typename C>
-struct Relocator<Stack<C> >;
-
 /**
  * Class of flexible-size generic stacks.
  * @since 11/03/2006 Bellevue
@@ -866,17 +863,6 @@ public:
   }
 
 };
-
-template<typename C>
-struct Relocator<Stack<C> >
-{
-  static void relocate(Stack<C>* oldStack, void* newAddr)
-  {
-    ::new(newAddr) Stack<C>(std::move(*oldStack));
-    oldStack->~Stack<C>();
-  }
-};
-
 
 } // namespace Lib
