+++ Lib/Timer.hpp
@@ -32,7 +32,6 @@ class Timer
 {
   Timer() : _running(false), _elapsed(0) { ensureTimerInitialized(); }
   ~Timer() { deinitializeTimer(); }
-  friend void ::checked_delete<Timer>(Timer*);
  
 public:
   CLASS_NAME(Timer);
