+++ Saturation/Splitter.hpp
@@ -99,7 +99,7 @@ private:
   Splitter& _parent;
 
   bool _solverIsSMT;
-  SATSolverSCP _solver;
+  ScopedPtr<SATSolver> _solver;
   ScopedPtr<DecisionProcedure> _dp;
   // use a separate copy of the decision procedure for ccModel computations and fill it up only with equalities
   ScopedPtr<SimpleCongruenceClosure> _dpModel;
