+++ Shell/Preprocess.cpp
@@ -57,7 +57,6 @@
 
 #include "UIHelper.hpp"
 #include "Lib/List.hpp"
-#include "Lib/RCPtr.hpp"
 
 #include "Kernel/TermIterators.hpp"
